replaced the entire concept of the last approach. There was an unefficient nested loop join between filehashes from mosaik2db-src and mosaik2db-dest.
Now all filehashes are sorted (external) and saved to file filehashes.idx. This allows to use a much more effectice merge join. About 50k mosaik2db is compared with itself in 50ms. The old approach needed several minutes.

The sorted filehashes are saved with their original offset in filehashes.bin. This is used to later print the right filenames of the duplicate files. Therefore the lookup index filenames.idx is used.